refactor(transform): isolate shaker statement graph
